Revert "Add database migrations via alembic"
