+++ docs/guide/concepts/infra.md
@@ -2,25 +2,6 @@
 
 !!! warning "Construction zone"
     This area of the docs is under active development and may change.
-
-## Databases
-### Sqlite
-By default, Marvin uses a Sqlite database located at `~/.marvin/marvin.sqlite`. You can customize this by setting `MARVIN_DATABASE_CONNECTION_URL` to `sqlite+aiosqlite:////{path/to/database}`.
-
-### Postgres
-Marvin can also use Postgres (though this isn't as actively tested at this time). To do so, install the postgres extra: `pip install marvin[postgres]` and set `MARVIN_DATABASE_CONNECTION_URL` to `postgresql+asyncpg://{username}:{password}@{hots}:{port}/{database}`, filling all variables appropriately.
-
-### Migrations
-Marvin keeps the database schema up-to-date with Alembic migrations. If Marvin detects an empty database, it will run the initial migration update automatically. However, subsequent migrations will not be run automatically (to avoid any conflicts). Instead, Marvin checks to see if the database is up-to-date on startup and prints a warning if it isn't. You can disable this behavior by setting `MARVIN_DATABASE_CHECK_MIGRATION_VERSION_ON_STARTUP=0`.
-
-After upgrading Marvin, or when you see the warning described above, you should upgrade the database by running:
-
-```shell
-marvin database upgrade
-```
-
-You will be asked to confirm the upgrade; pass `-y` to do so automatically (this can be useful in CI). The upgrade command is idempotent and safe to run multiple times; the database is only modified if necessary.
-
 ## Chroma 
 Marvin [provides a simple wrapper](https://github.com/PrefectHQ/marvin/blob/main/src/marvin/infra/chroma.py) of the ChromaDB client to make it easier to interact with the database.
