+++ pyproject.toml
@@ -7,7 +7,7 @@ description = "A batteries-included library for building AI-powered software."
 readme = "README.md"
 license = { file = "LICENSE" }
 classifiers = [
-    "Development Status :: 3 - Beta",
+    "Development Status :: 3 - Alpha",
     "Programming Language :: Python :: 3 :: Only",
     "Topic :: Scientific/Engineering :: Artificial Intelligence",
 ]
@@ -16,7 +16,6 @@ requires-python = ">=3.9"
 dependencies = [
     "aiofiles~=23.1",
     "aiosqlite~=0.18",
-    "alembic~=1.10",
     "asyncpg~=0.27",
     "cloudpickle~=2.2",
     "datamodel-code-generator~=0.17",
@@ -53,8 +52,6 @@ dependencies = [
 ]
 
 [project.optional-dependencies]
-
-
 dev = [
     "black[jupyter]",
     "ipython",
@@ -71,8 +68,8 @@ dev = [
     "pytest~=7.2",
     "ruff",
 ]
-postgres = ["asyncpg~=0.27.0"]
 chromadb = ["chromadb~=0.3"]
+
 
 [project.urls]
 Homepage = "https://github.com/prefecthq/marvin"
@@ -89,27 +86,20 @@ build-backend = "setuptools.build_meta"
 # pytest configuration
 [tool.pytest.ini_options]
 markers = ["llm: indicates that a test calls an LLM (may be slow)."]
-
-testpaths = ["tests"]
-
 norecursedirs = [
     "*.egg-info",
     ".git",
     ".mypy_cache",
-    ".pytest_cache",
-    ".ruff_cache",
-    ".vscode",
     "node_modules",
+    ".pytest_cache .vscode",
 ]
-
 asyncio_mode = 'auto'
-
 filterwarnings = [
     "ignore:'crypt' is deprecated and slated for removal in Python 3.13:DeprecationWarning",
 ]
 env = [
     "MARVIN_TEST_MODE=1",
-    "D:MARVIN_DATABASE_CONNECTION_URL=sqlite+aiosqlite:////tmp/marvin/marvin-tests.sqlite",
+    "D:MARVIN_DATABASE_CONNECTION_URL=sqlite+aiosqlite:////tmp/test.sqlite",
     "MARVIN_LOG_CONSOLE_WIDTH=120",
     "MARVIN_LOG_LEVEL=DEBUG",
     # disable auto profile picture creation
@@ -118,10 +108,8 @@ env = [
     'MARVIN_LOAD_BOT_DEFAULT_PLUGINS=0',
     'MARVIN_OPENAI_MODEL_TEMPERATURE=0',
     # use 3.5 for tests by default
-    'MARVIN_OPENAI_MODEL_NAME=gpt-3.5-turbo',
-    # don't create database on startup
-    'MARVIN_DATABASE_CHECK_MIGRATION_VERSION_ON_STARTUP=0'
-
+    'MARVIN_OPENAI_MODEL_NAME=gpt-3.5-turbo'
+    
 ]
 
 # black configuration
