+++ src/marvin/config.py
@@ -106,7 +106,6 @@ class Settings(BaseSettings):
     database_connection_url: SecretStr = (
         "sqlite+aiosqlite:////$MARVIN_HOME/marvin.sqlite"
     )
-    database_check_migration_version_on_startup: bool = True
 
     # GITHUB
     GITHUB_TOKEN: SecretStr = Field("", env=["MARVIN_GITHUB_TOKEN", "GITHUB_TOKEN"])
