+++ src/marvin/infra/__init__.py
@@ -1,1 +1,1 @@
-from . import database, chroma
+from . import db, chroma
