+++ src/marvin/models/threads.py
@@ -6,9 +6,9 @@ import sqlalchemy as sa
 from sqlmodel import Field
 
 import marvin
-from marvin.infra.database import AsyncSession, JSONType, provide_session
+from marvin.infra.db import AsyncSession, JSONType, provide_session
 from marvin.models.base import BaseSQLModel
-from marvin.models.ids import BotID, MessageID, ThreadID
+from marvin.models.ids import BotID, GenericID, MessageID, ThreadID
 from marvin.utilities.models import MarvinSQLModel
 from marvin.utilities.types import MarvinBaseModel
 
@@ -31,7 +31,7 @@ class Message(MarvinSQLModel, BaseMessage, table=True):
         sa.Index(
             "ix_message__thread_id_timestamp",
             "thread_id",
-            "timestamp",
+            sa.text("timestamp DESC"),
         ),
     )
     id: MessageID = Field(default_factory=MessageID.new, primary_key=True)
@@ -136,18 +136,22 @@ class ThreadRead(MarvinBaseModel):
         )
 
 
-# class ThreadSummary(MarvinSQLModel, table=True):
-#     """
-#     Table for storing summaries of threads at a certain point, to assist with
-#     long-term memory
-#     """
+class ThreadSummary(MarvinSQLModel, table=True):
+    """
+    Table for storing summaries of threads at a certain point, to assist with
+    long-term memory
+    """
 
-#     __table_args__ = (
-#         sa.ForeignKeyConstraint(["thread_id"], ["thread.id"], ondelete="CASCADE"),
-#         sa.Index("ix_thread_summary__thread_id_timestamp", "thread_id", "timestamp"),
-#     )
+    __table_args__ = (
+        sa.ForeignKeyConstraint(["thread_id"], ["thread.id"], ondelete="CASCADE"),
+        sa.Index(
+            "ix_thread_summary__thread_id_timestamp",
+            "thread_id",
+            sa.text("timestamp DESC"),
+        ),
+    )
 
-#     id: GenericID = Field(default_factory=GenericID.new, primary_key=True)
-#     thread_id: ThreadID
-#     summary: str
-#     timestamp: datetime.datetime
+    id: GenericID = Field(default_factory=GenericID.new, primary_key=True)
+    thread_id: ThreadID
+    summary: str
+    timestamp: datetime.datetime
